+++ mouse_controls.py
@@ -4,13 +4,11 @@ import time
 def left_mouse_down():
     win32api.mouse_event(win32con.MOUSEEVENTF_LEFTDOWN,0,0)
     time.sleep(.1)
-    print 'left Down'
 
 
 def left_mouse_up():
     win32api.mouse_event(win32con.MOUSEEVENTF_LEFTUP,0,0)
     time.sleep(.1)
-    print 'left release'
 
 
 def left_click():