+++ CHANGELOG.md
@@ -4,3 +4,6 @@
 - Always targets middle of the table.
 - Strength is always the same. (200px mouse offset)
 
+## v2
+- Image analysis
+- Targeting the closest ball.